refactor(npm): move lockfile and purl domain into project.py
Resolves https://github.com/hermetoproject/hermeto/issues/1156

Move Package, PackageLock, _Purlifier, NpmComponentInfo, and
ResolvedNpmPackage into npm/project.py.

Update npm/main.py to import project-domain symbols from
npm/project.py while keeping behavior unchanged.

No behavior changes.
